Remove global didCheckForGlobalCLIWarning and update debounce to use leading
