+++ packages/app/src/cli/services/dev/app-events/file-watcher.ts
@@ -69,9 +69,9 @@ export async function startFileWatcher(
 
   /**
    * Debounced function to emit the accumulated events.
-   * This function will be called at most once every 200ms to avoid emitting too many events in a short period.
+   * This function will be called at most once every 300ms to avoid emitting too many events in a short period.
    */
-  const debouncedEmit = debounce(emitEvents, 200)
+  const debouncedEmit = debounce(emitEvents, 300, {leading: true, trailing: true})
 
   /**
    * Emits the accumulated events and resets the current events list.
